nrt: move utility code outside plugin.go
plugin.go should contain only entry point and
orchestration code. Let's move all the utilties and logic
to other source code files.

Trivial code movement with minimal renames.
